+++ src/contexts/GroupsContextDispatch.jsx
@@ -1,79 +1,8 @@
-import React, { createContext, useContext, useReducer } from 'react';
+import React, { createContext, useReducer } from 'react';
 import PropTypes from 'prop-types';
+import groupReducer from '../reducers/group';
 
 const GroupsDispatchContext = createContext(undefined);
-
-function groupReducer(state, action) {
-  const { index } = action.payload;
-  switch (action.type) {
-    case 'setIndex':
-      return {
-        ...state,
-        index: action.payload,
-      };
-    case 'setImages':
-      return {
-        ...state,
-        images: action.payload,
-      };
-    case 'addGroup':
-      // Side note: index is here is optional. If passed, this will cause the
-      // state to update the index which will in turn set the selected group
-      // to the value of index. If index isn't passed, the selected group
-      // will remain unchanged.
-      return {
-        ...state,
-        groups: [...state.groups, action.payload.group],
-        index: index === null || index === undefined ? state.index : index,
-      };
-    case 'replaceGroups':
-      return {
-        ...state,
-        groups: action.payload.groups,
-        index: action.payload.index,
-      };
-    case 'addImage':
-      return {
-        ...state,
-        images: [...state.images, action.payload],
-      };
-    case 'init':
-      return {
-        groups: action.payload.groups,
-        images: action.payload.images,
-        index: action.payload.index,
-      };
-    case 'updateGroupMemberCount':
-      return {
-        ...state,
-        groups: action.payload,
-      };
-    default:
-      throw new Error(`Invalid Group reducer action ${action}.`);
-  }
-}
-
-/**
- * Fetches the group dispatch and state from the context.
- */
-function useGroups() {
-  const context = useContext(GroupsDispatchContext);
-
-  if (!context) {
-    throw new Error(
-      'Invalid useGroups hook, hook is not within the correct context.',
-    );
-  }
-
-  return context;
-}
-
-/**
- * Fetches the group state from the context.
- */
-function useGroupsState() {
-  return useGroups().state;
-}
 
 function GroupsProvider({ children }) {
   // Using an initial value of -1 here so that groupData can
@@ -98,8 +27,5 @@ GroupsProvider.propTypes = {
 
 export {
   GroupsDispatchContext as default,
-  groupReducer,
-  useGroups,
-  useGroupsState,
   GroupsProvider,
 };
