+++ src/contexts/LoadingContext.jsx
@@ -1,20 +1,10 @@
-import React, { createContext, useContext } from 'react';
+import React, { createContext } from 'react';
 import PropTypes from 'prop-types';
 
 const LoadingContext = createContext({
   groupLoading: false,
   imagesLoading: false,
 });
-
-function useLoading() {
-  const loading = useContext(LoadingContext);
-
-  if (!loading) {
-    throw new Error('Invalid useLoading hook, hook is not within the correct context.');
-  }
-
-  return loading;
-}
 
 function LoadingProvider({ value, children }) {
   return (
@@ -32,4 +22,4 @@ LoadingProvider.propTypes = {
   children: PropTypes.node.isRequired,
 };
 
-export { LoadingContext as default, useLoading, LoadingProvider };
+export { LoadingContext as default, LoadingProvider };
